Reflection (#12)
* reflection preview + utility anytuple

* keep working on serialization v3

* keep working on reflection

* reflection: tuple, istream, ostream

* reflection: keep working

* reflection: finalizing the 3rd gen mprpc protcol interface

* reflection: keep working on typemap

* reflection: keep working on serialization v3

* reflection: keep working on serialization v3

* reflection: serialization v3 tests pass + basic mprpc v3 test pass

* reflection: more refactoring and porting mprpc tests to proto v3

* reflection: rm serialization/v1, ported tests, examples and tutorials to v3

* reflection: ostream_reflection + tests + fixes for tests using future and exceptions

* reflection: passing tests on windows

* reflection: working on fixing compilation on macos

* reflection: fix some warnings on windows

* reflection: use PUBLIC in target_link_libraries for libraries

* reflection: clang format

* reflection: fix warning in binarydeserializer

+++ solid/frame/file/src/filestream.cpp
@@ -127,7 +127,7 @@ void FileBuf::device(FilePointerT& _rptr)
 
 ssize_t FileBuf::writeAll(const char* _s, size_t _n)
 {
-    int     wcnt = 0;
+    ssize_t wcnt = 0;
     ssize_t rv;
     do {
         rv = dev->write(_s, _n, off);
