+++ solid/frame/mprpc/src/mprpcmessagereader.cpp
@@ -230,7 +230,7 @@ const char* MessageReader::doConsumeMessage(
             if (message_size <= static_cast<size_t>(_pbufend - _pbufpos)) {
                 _receiver.context().pmessage_header = &rmsgstub.message_header_;
 
-                const long rv = rmsgstub.state_ == MessageStub::StateE::ReadHeadStart ? rmsgstub.deserializer_ptr_->run(_receiver.context(), _pbufpos, message_size, rmsgstub.message_header_) : rmsgstub.deserializer_ptr_->run(_receiver.context(), _pbufpos, message_size);
+                const ptrdiff_t rv = rmsgstub.state_ == MessageStub::StateE::ReadHeadStart ? rmsgstub.deserializer_ptr_->run(_receiver.context(), _pbufpos, message_size, rmsgstub.message_header_) : rmsgstub.deserializer_ptr_->run(_receiver.context(), _pbufpos, message_size);
 
                 rmsgstub.state_ = MessageStub::StateE::ReadHeadContinue;
                 _pbufpos += message_size;
@@ -297,7 +297,7 @@ const char* MessageReader::doConsumeMessage(
 
                     _receiver.context().pmessage_header = &rmsgstub.message_header_;
 
-                    const long rv = rmsgstub.state_ == MessageStub::StateE::ReadBodyStart ? rmsgstub.deserializer_ptr_->run(_receiver.context(), _pbufpos, message_size, rmsgstub.message_ptr_) : rmsgstub.deserializer_ptr_->run(_receiver.context(), _pbufpos, message_size);
+                    const ptrdiff_t rv = rmsgstub.state_ == MessageStub::StateE::ReadBodyStart ? rmsgstub.deserializer_ptr_->run(_receiver.context(), _pbufpos, message_size, rmsgstub.message_ptr_) : rmsgstub.deserializer_ptr_->run(_receiver.context(), _pbufpos, message_size);
 
                     rmsgstub.state_ = MessageStub::StateE::ReadBodyContinue;
                     _pbufpos += message_size;
